docs(PPDSC-2204): restructure dirs
