+++ site/components/release-notes/__tests__/release-badges.test.tsx
